refactor: move MaterialDefaultValueSetter into correct directory
